Remove BLOCKMAP limit

+++ src/doom/p_local.h
@@ -249,8 +249,8 @@ P_RadiusAttack
 // P_SETUP
 //
 extern byte*		rejectmatrix;	// for fast sight rejection
-extern short*		blockmaplump;	// offsets in blockmap are from here
-extern short*		blockmap;
+extern long*		blockmaplump;	// offsets in blockmap are from here
+extern long*		blockmap;
 extern int		bmapwidth;
 extern int		bmapheight;	// in mapblocks
 extern fixed_t		bmaporgx;
